Fix on displayed motherboard temperature. Fix on LBA28 last sector limit. Possible fix on CPU frequency calculation. More experimentation needed.

+++ drivers/cpu/cpu.h
@@ -27,6 +27,6 @@ extern void intel_interrupts_on(void);
 extern void cache_disable(void);
 extern void cache_enable(void);
 extern void cpuid(int op, int *eax, int *ebx, int *ecx, int *edx);
-extern unsigned long getCPUFreq(void);
+extern double getCPUFreq(void);
 void display_cpuid_update_microcode(void);
 void setup_ioapic(void);
